warning clear

+++ mc2/moe_distribute_buffer_reset/op_api/aclnn_moe_distribute_buffer_reset.cpp
@@ -67,7 +67,7 @@ static aclnnStatus CheckParams(const aclTensor *elasticInfo, const char *groupEp
 }
 
 aclnnStatus aclnnMoeDistributeBufferResetGetWorkspaceSize(const aclTensor *elasticInfo, const char *groupEp,
-                                                          int64_t epWorldSize, int64_t needSync,
+                                                          int32_t epWorldSize, int32_t needSync,
                                                           uint64_t *workspaceSize, aclOpExecutor **executor)
 
 {
